requirements.txt
numpy==1.23.4
pandas==1.5.1
Pillow==9.1.0
plotly==5.10.0
requests==2.28.1
streamlit==1.14.0
streamlit-extras==0.2.0
streamlit-timeline==0.0.2

